# src/main/go/projects/message-api/goa/gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"simple example api","description":"example api with hello world message service","version":"0.0.1"},"servers":[{"url":"http://0.0.0.0:8081","description":"hosts the message service"}],"paths":{"/messages":{"get":{"tags":["message"],"summary":"list message","operationId":"message#list","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Message"},"example":[{"id":"Amet officia dolor sit dolorem.","text":"Non tempore velit dolore dolorem dicta."},{"id":"Amet officia dolor sit dolorem.","text":"Non tempore velit dolore dolorem dicta."},{"id":"Amet officia dolor sit dolorem.","text":"Non tempore velit dolore dolorem dicta."}]},"example":[{"id":"Amet officia dolor sit dolorem.","text":"Non tempore velit dolore dolorem dicta."},{"id":"Amet officia dolor sit dolorem.","text":"Non tempore velit dolore dolorem dicta."}]}}}}},"post":{"tags":["message"],"summary":"create message","operationId":"message#create","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRequestBody"},"example":{"text":"a"}}}},"responses":{"201":{"description":"Created response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"},"example":{"id":"Eum nostrum voluptas illum.","text":"Est illo est sed ratione itaque."}}}}}}},"/messages/{id}":{"get":{"tags":["message"],"summary":"read message","operationId":"message#read","parameters":[{"name":"id","in":"path","description":"UUID of the message","required":true,"schema":{"type":"string","description":"UUID of the message","example":"Eveniet fuga."},"example":"Aspernatur dicta labore iste et iusto beatae."}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"},"example":{"id":"Cupiditate magni animi.","text":"Error veniam libero omnis odit."}}}},"404":{"description":"not_found: Not Found response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundError"},"example":{"message":"Id nemo eum beatae provident corporis."}}}}}}},"/openapi.json":{"get":{"tags":["message"],"summary":"Download gen/http/openapi3.json","operationId":"message#/openapi.json","responses":{"200":{"description":"File downloaded"}}}}},"components":{"schemas":{"CreateRequestBody":{"type":"object","properties":{"text":{"type":"string","description":"Message text","example":"pzz","minLength":1}},"example":{"text":"xo"},"required":["text"]},"Message":{"type":"object","properties":{"id":{"type":"string","description":"UUID of the message","example":"Occaecati ut itaque."},"text":{"type":"string","description":"Message text","example":"Labore cum sit illum suscipit."}},"example":{"id":"Quidem asperiores.","text":"Adipisci quod."},"required":["id","text"]},"NotFoundError":{"type":"object","properties":{"message":{"type":"string","description":"error message","example":"Ut et."}},"example":{"message":"Modi officia inventore aut fuga."},"required":["message"]}}},"tags":[{"name":"message","description":"simple message service"}]}